client and server setup
