+++ client/src/components/ButtonAppBar.js
@@ -1,0 +1,38 @@
+import * as React from 'react';
+import AppBar from '@mui/material/AppBar';
+import Box from '@mui/material/Box';
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography';
+import Button from '@mui/material/Button';
+import IconButton from '@mui/material/IconButton';
+import MenuIcon from '@mui/icons-material/Menu';
+// import { color } from '@mui/system';
+// import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+
+import './style/ButtonAppBar.css'
+
+export default function ButtonAppBar() {
+  return (
+    <Box sx={{ flexGrow: 1 }}>
+      <AppBar position="static">
+        <Toolbar>
+          <IconButton
+            size="large"
+            edge="start"
+            color="inherit"
+            aria-label="menu"
+            sx={{ mr: 2 }}
+          >
+            <MenuIcon />
+          </IconButton>
+          <Typography variant="h6" component="div" sx={{ flexGrow: 1 }}>
+            Yash Chat App
+          </Typography>
+          <Button className='signupbtn' color='inherit'>Signup</Button>
+          <Button className='signupbtn' color='inherit'>Login</Button>
+
+        </Toolbar>
+      </AppBar>
+    </Box>
+  );
+}
