+++ client/src/components/style/ButtonAppBar.css
@@ -1,0 +1,3 @@
+.signupbtn{
+    background-color: #4caf50;
+}
