--- index.html
@@ -1,63 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Chatroom App</title>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-    <style>
-        .note{
-          margin-top: 100px;
-      
-          color: red;
-        }
-    </style>
-    <script>
-        function passv(){
-            var firstname = document.getElementById("firstname").value;
-            var password = document.getElementById("password").value;
-            
-            localStorage.setItem("textvalue",firstname);
-
-            
-            
-            
-            return false;
-        }
-    </script>
-</head>
-<body>
-    <!--
-    <form action="main.html">
-        <input type="text" id="txt">
-        <button type="submit" onclick="passv()">submit</button>
-    </form>---->
-    
-    <center>
-      <div class="note"><b>Note :- </b>currently password function is disabled, you may just click button to sign in.</div>
-    
-    <br><br>
-    <form action="main.html">
-        <div class="form-group row">
-          <label for="inputEmail3" class="col-sm-2 col-form-label">Username</label>
-          <div class="col-sm-10">
-            <input type="text" class="form-control" id="firstname" placeholder="Username">
-          </div>
-        </div>
-        <div class="form-group row">
-          <label for="inputPassword3" class="col-sm-2 col-form-label">Password</label>
-          <div class="col-sm-10">
-            <input type="password" class="form-control" id="password" placeholder="Password">
-          </div>
-        </div>
-        <div class="form-group row">
-          <div class="col-sm-10">
-            <button type="submit" class="btn btn-primary" onclick="passv()">Sign in</button>
-          </div>
-        </div>
-      </form>
-    </center>
-
-</body>
-</html>
