+++ old/firebase-debug.log
@@ -1,0 +1,66 @@
+[debug] [2022-06-09T03:20:36.523Z] ----------------------------------------------------------------------
+[debug] [2022-06-09T03:20:36.526Z] Command:       F:\nodejs1\node.exe C:\Users\HP\AppData\Roaming\npm\node_modules\firebase-tools\lib\bin\firebase.js login
+[debug] [2022-06-09T03:20:36.527Z] CLI Version:   11.0.1
+[debug] [2022-06-09T03:20:36.527Z] Platform:      win32
+[debug] [2022-06-09T03:20:36.528Z] Node Version:  v16.15.0
+[debug] [2022-06-09T03:20:36.529Z] Time:          Thu Jun 09 2022 08:50:36 GMT+0530 (India Standard Time)
+[debug] [2022-06-09T03:20:36.529Z] ----------------------------------------------------------------------
+[debug] 
+[debug] [2022-06-09T03:20:36.533Z] >>> [apiv2][query] GET https://firebase-public.firebaseio.com/cli.json [none]
+[info] i  Firebase optionally collects CLI usage and error reporting information to help improve our products. Data is collected in accordance with Google's privacy policy (https://policies.google.com/privacy) and is not used to identify you.
+ 
+[debug] [2022-06-09T03:20:37.528Z] <<< [apiv2][status] GET https://firebase-public.firebaseio.com/cli.json 200
+[debug] [2022-06-09T03:20:37.529Z] <<< [apiv2][body] GET https://firebase-public.firebaseio.com/cli.json {"cloudBuildErrorAfter":1594252800000,"cloudBuildWarnAfter":1590019200000,"defaultNode10After":1594252800000,"minVersion":"3.0.5","node8DeploysDisabledAfter":1613390400000,"node8RuntimeDisabledAfter":1615809600000,"node8WarnAfter":1600128000000}
+[info] i  To change your data collection preference at any time, run `firebase logout` and log in again. 
+[info] 
+[info] Visit this URL on this device to log in:
+[info] https://accounts.google.com/o/oauth2/auth?client_id=563584335869-fgrhgmd47bqnekij5i8b5pr03ho849e6.apps.googleusercontent.com&scope=email%20openid%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fcloudplatformprojects.readonly%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Ffirebase%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fcloud-platform&response_type=code&state=348497685&redirect_uri=http%3A%2F%2Flocalhost%3A9005
+[info] 
+[info] Waiting for authentication...
+[debug] [2022-06-09T03:21:30.321Z] ----------------------------------------------------------------------
+[debug] [2022-06-09T03:21:30.324Z] Command:       F:\nodejs1\node.exe C:\Users\HP\AppData\Roaming\npm\node_modules\firebase-tools\lib\bin\firebase.js login
+[debug] [2022-06-09T03:21:30.324Z] CLI Version:   11.0.1
+[debug] [2022-06-09T03:21:30.325Z] Platform:      win32
+[debug] [2022-06-09T03:21:30.325Z] Node Version:  v16.15.0
+[debug] [2022-06-09T03:21:30.326Z] Time:          Thu Jun 09 2022 08:51:30 GMT+0530 (India Standard Time)
+[debug] [2022-06-09T03:21:30.326Z] ----------------------------------------------------------------------
+[debug] 
+[info] i  Firebase optionally collects CLI usage and error reporting information to help improve our products. Data is collected in accordance with Google's privacy policy (https://policies.google.com/privacy) and is not used to identify you.
+ 
+[info] 
+[info] Visit this URL on this device to log in:
+[info] https://accounts.google.com/o/oauth2/auth?client_id=563584335869-fgrhgmd47bqnekij5i8b5pr03ho849e6.apps.googleusercontent.com&scope=email%20openid%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fcloudplatformprojects.readonly%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Ffirebase%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fcloud-platform&response_type=code&state=938022680&redirect_uri=http%3A%2F%2Flocalhost%3A9005
+[info] 
+[info] Waiting for authentication...
+[debug] [2022-06-09T03:22:09.047Z] >>> [apiv2][query] POST https://auth.firebase.tools/attest [none]
+[debug] [2022-06-09T03:22:09.048Z] >>> [apiv2][body] POST https://auth.firebase.tools/attest {"session_id":"6570abfa-208b-46a1-8c9a-9e8b0d66bed2"}
+[debug] [2022-06-09T03:22:10.898Z] <<< [apiv2][status] POST https://auth.firebase.tools/attest 200
+[debug] [2022-06-09T03:22:10.899Z] <<< [apiv2][body] POST https://auth.firebase.tools/attest {"token":"_n33kkSZ6NJy_H3LHNxPn1LvkmP7Vr75rhWbFLpWIZg"}
+[info] 
+[info] To sign in to the Firebase CLI:
+[info] 
+[info] 1. Take note of your session ID:
+[info] 
+[info]    6570A
+[info] 
+[info] 2. Visit the URL below on any device and follow the instructions to get your code:
+[info] 
+[info]    https://auth.firebase.tools/login?code_challenge=ZiovHj5yjjfLDZQcv5t0hZ4ojqZ6eqT17DVOwmcNEMI&session=6570abfa-208b-46a1-8c9a-9e8b0d66bed2&attest=_n33kkSZ6NJy_H3LHNxPn1LvkmP7Vr75rhWbFLpWIZg
+[info] 
+[info] 3. Paste or enter the authorization code below once you have it:
+[info] 
+[debug] [2022-06-09T03:22:27.299Z] ----------------------------------------------------------------------
+[debug] [2022-06-09T03:22:27.303Z] Command:       F:\nodejs1\node.exe C:\Users\HP\AppData\Roaming\npm\node_modules\firebase-tools\lib\bin\firebase.js login
+[debug] [2022-06-09T03:22:27.304Z] CLI Version:   11.0.1
+[debug] [2022-06-09T03:22:27.304Z] Platform:      win32
+[debug] [2022-06-09T03:22:27.304Z] Node Version:  v16.15.0
+[debug] [2022-06-09T03:22:27.305Z] Time:          Thu Jun 09 2022 08:52:27 GMT+0530 (India Standard Time)
+[debug] [2022-06-09T03:22:27.306Z] ----------------------------------------------------------------------
+[debug] 
+[info] i  Firebase optionally collects CLI usage and error reporting information to help improve our products. Data is collected in accordance with Google's privacy policy (https://policies.google.com/privacy) and is not used to identify you.
+ 
+[info] 
+[info] Visit this URL on this device to log in:
+[info] https://accounts.google.com/o/oauth2/auth?client_id=563584335869-fgrhgmd47bqnekij5i8b5pr03ho849e6.apps.googleusercontent.com&scope=email%20openid%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fcloudplatformprojects.readonly%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Ffirebase%20https%3A%2F%2Fwww.googleapis.com%2Fauth%2Fcloud-platform&response_type=code&state=450700944&redirect_uri=http%3A%2F%2Flocalhost%3A9005
+[info] 
+[info] Waiting for authentication...
