--- style.css
@@ -1,82 +1,0 @@
-h1{
-    
-    margin: auto;
-    color: rgb(0, 166, 255);
-}
-
-
-#box{
-    max-width: 955px;
-    border: 2px solid black;
-    margin: auto;
-    height: 60vh;
-    padding: 33px;
-    overflow-y: scroll;
-    margin-bottom: 23px;
-
-}
-
-#send-container{
-    display: block;
-    margin: auto;
-    text-align: center;
-    max-width: 985px;
-    width: 100%;
-
-
-}
-#message{
-    width: 70%;
-    border: 2px solid black;
-    border-radius: 6px;
-    height: 34px;
-
-}
-#send-button{
-    cursor: pointer;
-    border: 2px solid black;
-    border-radius: 6px;
-    height: 34px;
-}
-
-.message{
-    background-color: grey;
-    width: 24%;
-    padding: 10px;
-    margin: 17px 2px;
-    border: 2px solid black;
-    border-radius: 10px;
-    overflow-y: auto;
-
-
-
-}
-.left{
-    background-color: grey;
-    width: 24%;
-    padding: 10px;
-    margin: 17px 2px;
-    border: 2px solid black;
-    border-radius: 10px;
-    overflow-y: auto;
-    float: left;
-    clear: both;
-
-}
-.right{
-    background-color: grey;
-    width: 24%;
-    padding: 10px;
-    margin: 17px 2px;
-    border: 2px solid black;
-    border-radius: 10px;
-    float: right;
-    clear: both;
-    overflow-y: auto;
-}
-footer{
-    padding: 50px;
-    background-color: black;
-    color: white;
-    height: 500px;
-}
